+++ src/JT809.Protocol/SubMessageBody/JT809_0x9200_0x9202.cs
@@ -26,28 +26,56 @@ namespace JT809.Protocol.SubMessageBody
 
         public override string Description => "交换车辆定位信息消息";
         /// <summary>
-        /// 车辆定位信息
+        /// 车辆定位信息 2011版本
         /// </summary>
-        public VehiclePositionPropertieBase VehiclePosition { get; set; }
+        public JT809VehiclePositionProperties VehiclePosition { get; set; }
+        /// <summary>
+        /// 车辆定位信息 2019版本
+        /// </summary>
+        public JT809VehiclePositionProperties_2019 VehiclePosition_2019 { get; set; }
 
         public void Analyze(ref JT809MessagePackReader reader, Utf8JsonWriter writer, IJT809Config config)
         {
-            JT809_0x9200_0x9202 value = new JT809_0x9200_0x9202();
-            writer.WriteStartObject("车辆定位信息");
-            VehiclePositionPropertieBase.Analyze(ref reader, writer, config);
-            writer.WriteEndObject();
+            if (config.Version == JT809Version.JTT2011)
+            {
+                config.GetMessagePackFormatter<JT809VehiclePositionProperties>().Analyze(ref reader, writer, config);
+            }
+            else if (config.Version == JT809Version.JTT2019)
+            {
+                config.GetMessagePackFormatter<JT809VehiclePositionProperties_2019>().Analyze(ref reader, writer, config);
+            }
         }
 
         public JT809_0x9200_0x9202 Deserialize(ref JT809MessagePackReader reader, IJT809Config config)
         {
             JT809_0x9200_0x9202 value = new JT809_0x9200_0x9202();
-            value.VehiclePosition = VehiclePositionPropertieBase.Deserialize(ref reader, config);
+            if (config.Version == JT809Version.JTT2011)
+            {
+                value.VehiclePosition = config.GetMessagePackFormatter<JT809VehiclePositionProperties>().Deserialize(ref reader, config);
+            }
+            else if (config.Version == JT809Version.JTT2019)
+            {
+                value.VehiclePosition_2019 = config.GetMessagePackFormatter<JT809VehiclePositionProperties_2019>().Deserialize(ref reader, config);
+            }
             return value;
         }
 
         public void Serialize(ref JT809MessagePackWriter writer, JT809_0x9200_0x9202 value, IJT809Config config)
         {
-            VehiclePositionPropertieBase.Serialize(ref writer, value.VehiclePosition, config);
+            if (config.Version == JT809Version.JTT2011)
+            {
+                if (value.VehiclePosition != null)
+                {
+                    config.GetMessagePackFormatter<JT809VehiclePositionProperties>().Serialize(ref writer, value.VehiclePosition, config);
+                }
+            }
+            else if (config.Version == JT809Version.JTT2019)
+            {
+                if (value.VehiclePosition_2019 != null)
+                {
+                    config.GetMessagePackFormatter<JT809VehiclePositionProperties_2019>().Serialize(ref writer, value.VehiclePosition_2019, config);
+                }
+            }
         }
     }
 }
